+++ src/main/java/coupon/coupon/domain/CouponDiscountAmount.java
@@ -1,4 +1,4 @@
-package coupon.domain.coupon;
+package coupon.coupon.domain;
 
 import lombok.Getter;
