+++ src/main/java/coupon/coupon/domain/CouponMinimumOrderAmount.java
@@ -1,4 +1,4 @@
-package coupon.domain.coupon;
+package coupon.coupon.domain;
 
 import lombok.Getter;
