+++ src/main/java/coupon/coupon/domain/IssuancePeriod.java
@@ -1,4 +1,4 @@
-package coupon.domain.coupon;
+package coupon.coupon.domain;
 
 import java.time.LocalDateTime;
