+++ src/main/java/coupon/coupon/domain/ProductionCategory.java
@@ -1,4 +1,4 @@
-package coupon.domain.coupon;
+package coupon.coupon.domain;
 
 public enum ProductionCategory {
     FASHION, HOME_APPLIANCES, FURNITURE, FOOD
