+++ src/main/java/coupon/coupon/infrastructure/repository/CouponJpaRepository.java
@@ -1,4 +1,4 @@
-package coupon.infrastructure.repository;
+package coupon.coupon.infrastructure.repository;
 
 import java.util.Optional;
