+++ src/test/java/coupon/domain/coupon/CouponMinimumOrderAmountTest.java
@@ -8,6 +8,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
+import coupon.coupon.domain.CouponMinimumOrderAmount;
+
 class CouponMinimumOrderAmountTest {
 
     @DisplayName("유효한 최소 주문 금액이 입력되면 객체를 생성한다.")
